backend/requirements.txt
Flask
pymongo[srv]
Flask-PyMongo
flask_cors
requests
pandas
langchain
python-dotenv